+++ src/layers/GeomodelLayer.ts
@@ -11,6 +11,7 @@ import { default as curveCatmullRom } from 'cat-rom-spline';
 
 class GeomodelLayer extends WebGLLayer {
   options: GeomodelLayerOptions;
+  curveOptions = { samples: 50, knot: 0.5 };
 
   constructor(id: String, options: GeomodelLayerOptions) {
     super(id, options);
@@ -31,82 +32,32 @@ class GeomodelLayer extends WebGLLayer {
   }
 
   render(event: OnRescaleEvent | OnUpdateEvent) {
-    const { xscale, yscale } = event;
-    const [, width] = xscale.range();
-    const [, height] = yscale.range();
-
-    console.log(xscale.domain());
-
-    const data: GeoModelData[] = [
-      {
-        name: 'strat 1',
-        color: 0xff000,
-        md: [70, 90, 100, 110, 100, 100],
-        pos: [
-          [0, 99],
-          [100, 99],
-          [200, 99],
-          [450, 99],
-          [600, 99],
-          [1000, 99],
-        ],
-      },
-      {
-        name: 'strat 2',
-        color: 0xffff0,
-        md: [100 + 50, 120 + 50, 100 + 50, 140 + 50, 100 + 50, 120 + 50],
-        pos: [
-          [0, 99],
-          [140, 99],
-          [200, 99],
-          [400, 99],
-          [750, 99],
-          [1000, 99],
-        ],
-      },
-      {
-        name: 'strat 3',
-        color: 0xff00ff,
-        md: [100 + 150, 120 + 120, 100 + 170, 140 + 150, 100 + 150, 177 + 150],
-        pos: [
-          [0, 99],
-          [190, 99],
-          [200, 99],
-          [520, 99],
-          [850, 99],
-          [1000, 99],
-        ],
-      },
-      {
-        name: 'strat 4',
-        color: 0x0330ff,
-        md: [100 + 250, 120 + 220, 100 + 270, 100 + 250, 140 + 250, 177 + 250],
-        pos: [
-          [0, 99],
-          [200, 99],
-          [210, 99],
-          [300, 99],
-          [750, 99],
-          [1000, 99],
-        ],
-      },
-    ];
-
+    const { xscale, yscale, data } = event;
     const scaledData = this.generateScaledData(data, xscale, yscale);
 
     // for each layer
     // paint from top down, with color
     // missing data is 0, dont color
 
-    data.forEach((s, index) => {
-      const area = new Graphics();
-      area.lineStyle(4, s.color, 1);
-      area.beginFill(s.color);
-      this.renderAreaTopLine(area, scaledData[index], xscale, yscale);
-      area.endFill();
-      this.ctx.stage.addChild(area);
-    });
+    data.forEach((s: GeoModelData, index: number) =>
+      this.drawArea(s, index, scaledData, xscale, yscale),
+    );
   }
+
+  drawArea = (
+    s: GeoModelData,
+    index: number,
+    scaledData: [number, number][][],
+    xscale: ScaleLinear<number, number>,
+    yscale: ScaleLinear<number, number>,
+  ) => {
+    const area = new Graphics();
+    area.lineStyle(4, s.color, 1);
+    area.beginFill(s.color);
+    this.renderAreaTopLine(area, scaledData[index], xscale, yscale);
+    area.endFill();
+    this.ctx.stage.addChild(area);
+  };
 
   generateScaledData = (
     data: GeoModelData[],
@@ -133,29 +84,23 @@ class GeomodelLayer extends WebGLLayer {
     xscale: ScaleLinear<number, number>,
     yscale: ScaleLinear<number, number>,
   ): void => {
-    const points = data;
-    const options = { samples: 50, knot: 0.5 };
-
-    const interpolatedPoints = curveCatmullRom(points, options);
-    // .filter(
-    //   (x: [number, number]) => !isNaN(x[0]) && !isNaN(x[1]),
-    // );
-
+    const interpolatedPoints = curveCatmullRom(data, this.curveOptions);
     const [leftMostX, rightMostX] = xscale.range();
     const [, bottom] = yscale.range();
-
-    console.log('interpolatedPoints', points);
-    area.drawPolygon([
+    const topRightPoint = data[data.length - 1][1];
+    const topLeftPoint = data[0][1];
+    const points = [
       leftMostX,
       bottom,
       leftMostX,
-      points[0][1], // top left point
-      ...interpolatedPoints.flat(), // actual points
+      topLeftPoint,
+      ...interpolatedPoints.flat(),
       rightMostX,
-      points[points.length - 1][1], // top right point
+      topRightPoint,
       rightMostX,
       bottom,
-    ]);
+    ];
+    area.drawPolygon(points);
   };
 }
